WIP: vendor generated Go clients
We want to be able to test generated clients from this repo.
However, we added import comment saying these clients belong in
"cloud.google.com/go".
Work around this by putting all clients in vendor directory,
which disables the canonical import check.

Packages internal/version and internal/testutil are redeclared
in this repo.
Package testutil should later be changed to match how we plumb
project ID and credentials through CircleCI.

Some tests do not build yet, since they depend on testutil.UIDSpace.
This should be fixed by googleapis/toolkit#1279.

If this new layout looks good to reviewers,
a corresponding change to be made to googleapis to preserve the layout.
Hence, the WIP in the commit title.