create easier way to make javadocs
Introduce a new Gradle build target "allDocs" that creates Javadoc
for all proto- and grpc- subprojects.
This fixes a bug where grpc and proto docs weren't correctly linked
together.

A bash script, javadoc.sh, copies and commits the docs, preparing it for
push. The functionality can be put in a Gradle task but it seems like an
overkill for executing straight-line code.